+++ spec/slim_lint/ruby_extractor_spec.rb
@@ -1,13 +1,16 @@
 require 'spec_helper'
 
 describe SlimLint::RubyExtractor do
+  let(:extractor) { described_class.new }
+
   describe '#extract' do
     let(:sexp) { SlimLint::RubyExtractEngine.new.call(normalize_indent(slim)) }
-    subject { super().extract(sexp) }
+    subject { extractor.extract(sexp) }
 
     context 'with an empty Slim document' do
       let(:slim) { '' }
-      it { should == '' }
+      its(:source) { should == '' }
+      its(:source_map) { should == {} }
     end
 
     context 'with verbatim text' do
@@ -15,7 +18,8 @@ describe SlimLint::RubyExtractor do
         | Hello world
       SLIM
 
-      it { should == 'puts' }
+      its(:source) { should == 'puts' }
+      its(:source_map) { should == { 1 => 1 } }
     end
 
     context 'with verbatim text on multiple lines' do
@@ -25,7 +29,8 @@ describe SlimLint::RubyExtractor do
           world
       SLIM
 
-      it { should == 'puts' }
+      its(:source) { should == 'puts' }
+      its(:source_map) { should == { 1 => 2 } }
     end
 
     context 'with verbatim text with trailing whitespace' do
@@ -33,15 +38,17 @@ describe SlimLint::RubyExtractor do
         ' Hello world
       SLIM
 
-      it { should == 'puts' }
+      its(:source) { should == 'puts' }
+      its(:source_map) { should == { 1 => 1 } }
     end
 
-    context 'with inline HTML' do
+    context 'with inline static HTML' do
       let(:slim) { <<-SLIM }
         <p><b>Hello world!</b></p>
       SLIM
 
-      it { should == 'puts' }
+      its(:source) { should == 'puts' }
+      its(:source_map) { should == { 1 => 1 } }
     end
 
     context 'with control code' do
@@ -49,7 +56,8 @@ describe SlimLint::RubyExtractor do
         - some_expression
       SLIM
 
-      it { should == 'some_expression' }
+      its(:source) { should == 'some_expression' }
+      its(:source_map) { should == { 1 => 1 } }
     end
 
     context 'with output code' do
@@ -57,7 +65,23 @@ describe SlimLint::RubyExtractor do
         = some_expression
       SLIM
 
-      it { should == 'some_expression' }
+      its(:source) { should == 'some_expression' }
+      its(:source_map) { should == { 1 => 1 } }
+    end
+
+    context 'with output code with block contents' do
+      let(:slim) { <<-SLIM }
+        = simple_form_for User.new, url: some_url do |f|
+          = f.input :email, autofocus: true
+      SLIM
+
+      its(:source) { should == normalize_indent(<<-RUBY).chomp }
+        simple_form_for User.new, url: some_url do |f|
+        f.input :email, autofocus: true
+        end
+      RUBY
+
+      its(:source_map) { { 1 => 1, 2 => 2, 3 => 1 } }
     end
 
     context 'with output without escaping' do
@@ -65,7 +89,8 @@ describe SlimLint::RubyExtractor do
         == some_expression
       SLIM
 
-      it { should == 'some_expression' }
+      its(:source) { should == 'some_expression' }
+      its(:source_map) { should == { 1 => 1 } }
     end
 
     context 'with a code comment' do
@@ -73,7 +98,8 @@ describe SlimLint::RubyExtractor do
         / This line will not appear
       SLIM
 
-      it { should == '' }
+      its(:source) { should == '' }
+      its(:source_map) { should == {} }
     end
 
     context 'with an HTML comment' do
@@ -81,7 +107,8 @@ describe SlimLint::RubyExtractor do
         /! This line will appear
       SLIM
 
-      it { should == 'puts' }
+      its(:source) { should == 'puts' }
+      its(:source_map) { should == { 1 => 1 } }
     end
 
     context 'with an Internet Explorer conditional comment' do
@@ -90,7 +117,8 @@ describe SlimLint::RubyExtractor do
           Get a better browser
       SLIM
 
-      it { should == "puts\nputs" }
+      its(:source) { should == "puts\nputs" }
+      its(:source_map) { should == { 1 => 2, 2 => 2 } }
     end
 
     context 'with doctype tag' do
@@ -98,7 +126,8 @@ describe SlimLint::RubyExtractor do
         doctype xml
       SLIM
 
-      it { should == 'puts' }
+      its(:source) { should == 'puts' }
+      its(:source_map) { should == { 1 => 1 } }
     end
 
     context 'with an HTML tag' do
@@ -106,7 +135,8 @@ describe SlimLint::RubyExtractor do
         p A paragraph
       SLIM
 
-      it { should == "puts\nputs" }
+      its(:source) { should == "puts\nputs" }
+      its(:source_map) { should == { 1 => 1, 2 => 1 } }
     end
 
     context 'with an HTML tag with interpolation' do
@@ -114,7 +144,8 @@ describe SlimLint::RubyExtractor do
         p A \#{adjective} paragraph for \#{noun}
       SLIM
 
-      it { should == "puts\nputs\nadjective\nputs\nnoun" }
+      its(:source) { should == "puts\nputs\nadjective\nputs\nnoun" }
+      its(:source_map) { should == { 1 => 1, 2 => 1, 3 => 1, 4 => 1, 5 => 1 } }
     end
 
     context 'with an HTML tag with static attributes' do
@@ -122,7 +153,8 @@ describe SlimLint::RubyExtractor do
         p class="highlight"
       SLIM
 
-      it { should == "puts\nputs" }
+      its(:source) { should == "puts\nputs" }
+      its(:source_map) { should == { 1 => 1, 2 => 1 } }
     end
 
     context 'with an HTML tag with Ruby attributes' do
@@ -130,9 +162,13 @@ describe SlimLint::RubyExtractor do
         p class=user.class id=user.id
       SLIM
 
-      it { should include 'puts' }
-      it { should include 'user.class' }
-      it { should include 'user.id' }
+      its(:source) { should == normalize_indent(<<-RUBY).chomp }
+        puts
+        user.class
+        user.id
+      RUBY
+
+      its(:source_map) { should == { 1 => 1, 2 => 1, 3 => 1 } }
     end
 
     context 'with a dynamic tag splat' do
@@ -140,9 +176,13 @@ describe SlimLint::RubyExtractor do
         *some_dynamic_tag Hello World!
       SLIM
 
-      # Splatting injects some Slim helper code, so we need to relax the test
-      it { should include 'some_dynamic_tag' }
-      it { should include 'puts' }
+      its(:source) { should == normalize_indent(<<-RUBY).chomp }
+        puts
+        some_dynamic_tag
+        puts
+      RUBY
+
+      its(:source_map) { should == { 1 => 1, 2 => 1, 3 => 1 } }
     end
 
     context 'with an if statement' do
@@ -151,11 +191,13 @@ describe SlimLint::RubyExtractor do
           | It's true!
       SLIM
 
-      it { should == normalize_indent(<<-RUBY).chomp }
+      its(:source) { should == normalize_indent(<<-RUBY).chomp }
         if condition_true?
         puts
         end
       RUBY
+
+      its(:source_map) { should == { 1 => 1, 2 => 2, 3 => 3 } }
     end
 
     context 'with an if/else statement' do
@@ -166,13 +208,15 @@ describe SlimLint::RubyExtractor do
           | It's false!
       SLIM
 
-      it { should == normalize_indent(<<-RUBY).chomp }
+      its(:source) { should == normalize_indent(<<-RUBY).chomp }
         if condition_true?
         puts
         else
         puts
         end
       RUBY
+
+      its(:source_map) { should == { 1 => 1, 2 => 2, 3 => 3, 4 => 4, 5 => 5 } }
     end
 
     context 'with an if/elsif/else statement' do
@@ -185,7 +229,7 @@ describe SlimLint::RubyExtractor do
           | It's false!
       SLIM
 
-      it { should == normalize_indent(<<-RUBY).chomp }
+      its(:source) { should == normalize_indent(<<-RUBY).chomp }
         if condition_true?
         puts
         elsif something_else?
@@ -194,6 +238,79 @@ describe SlimLint::RubyExtractor do
         puts
         end
       RUBY
+
+      its(:source_map) { should == { 1 => 1, 2 => 2, 3 => 3, 4 => 4, 5 => 5, 6 => 6, 7 => 7 } }
+    end
+
+    context 'with an if/else statement with statements following it' do
+      let(:slim) { <<-SLIM }
+        - if condition_true?
+          | It's true!
+        - else
+          | It's false!
+        - following_statement
+        - another_statement
+      SLIM
+
+      its(:source) { should == normalize_indent(<<-RUBY).chomp }
+        if condition_true?
+        puts
+        else
+        puts
+        end
+        following_statement
+        another_statement
+      RUBY
+
+      its(:source_map) { should == { 1 => 1, 2 => 2, 3 => 3, 4 => 4, 5 => 5, 6 => 5, 7 => 6 } }
+    end
+
+    context 'with an output statement with statements following it' do
+      let(:slim) { <<-SLIM }
+        = some_output
+        - some_statement
+        - another_statement
+      SLIM
+
+      its(:source) { should == normalize_indent(<<-RUBY).chomp }
+        some_output
+        some_statement
+        another_statement
+      RUBY
+
+      its(:source_map) { should == { 1 => 1, 2 => 2, 3 => 3 } }
+    end
+
+    context 'with an output statement that spans multiple lines' do
+      let(:slim) { <<-SLIM }
+        = some_output 1,
+                      2,
+                      3
+      SLIM
+
+      its(:source) { should == normalize_indent(<<-RUBY).chomp }
+      some_output 1,
+      2,
+      3
+      RUBY
+
+      its(:source_map) { should == { 1 => 1, 2 => 2, 3 => 3 } }
+    end
+
+    context 'with a control statement that spans multiple lines' do
+      let(:slim) { <<-SLIM }
+        - some_method 1,
+                      2,
+                      3
+      SLIM
+
+      its(:source) { should == normalize_indent(<<-RUBY).chomp }
+      some_method 1,
+      2,
+      3
+      RUBY
+
+      its(:source_map) { should == { 1 => 1, 2 => 2, 3 => 3 } }
     end
   end
 end